# public/swagger.json
{"swagger":"2.0",
	"info":{
		"description":"These are the list of apis which is used to comunicate with the app.",
		"version":"1.0.0",
		"title":"Socialcops Interview",
		"termsOfService":"",
		"contact":{"email":""},
		"license":{}},
	"host":"localhost:3000",
	"basePath":"/api",
		"tags":[{
			"name":"Login",
			"description":"Authentication System"},
		{
			"name":"Patch json",
			"description":"Json patch functionality"},
		{
			"name":"Upload",
			"description":"Upload image",
			"externalDocs":{"description":"Upload image link here"}
		}],
		"schemes":["http"],
			"paths":{
				"/login":{
				"post":{
					"tags":["Login"],
					"summary":"Login into the system",
					"description":"",
					"operationId":"",
					"consumes":["application/json","application/xml"],
					"produces":["application/json"],
					"parameters":[{
						"in":"body",
						"name":"body",
						"description":"Username and passowrd credential",
						"required":true,
						"schema":{"$ref":"#/definitions/Pet"}}],
					"responses":{"405":{"description":"Invalid input"}},
					"security":[]}},
				"/json_patch":{
				"post":{
					"tags":["Patch json"],
					"summary":"Json Patch Functionality",
					"description":"",
					"operationId":"",
					"produces":["application/json"],
					"parameters":[{
						"name":"body",
						"in":"body",
						"description":"Status values that need to be considered for filter",
						"required":true,
						"collectionFormat":"multi",
						"schema":{"$ref":"#/definitions/Tag"}}],
					"responses":{
						"200":{
							"description":"successful operation"},
						"400":{"description":"Invalid status value"}},
					"security":[]}},
				"/upload":{
					"post":{
						"tags":["Upload"],
						"summary":"upload Image",
						"description":"",
						"operationId":"",
						"produces":["application/json"],
						"parameters":[{
							"name":"body",
							"in":"body",
							"description":"Tags to filter by",
							"required":true,
							"items":{"type":"string"},
							"schema":{"$ref":"#/definitions/User"},
							"collectionFormat":"multi"}],
					"responses":{
						"200":{"description":"successful operation"},
						"400":{"description":"Invalid tag value"}},
					"security":[]}}},
					"securityDefinitions":{
						"petstore_auth":{
							"type":"oauth2",
							"authorizationUrl":"http://petstore.swagger.io/oauth/dialog",
							"flow":"implicit",
							"scopes":{
								"write:pets":"modify pets in your account",
								"read:pets":"read your pets"}},
						"api_key":{
							"type":"apiKey",
							"name":"api_key",
							"in":"header"}},
						"definitions":{
							"Order":{
								"type":"object",
								"properties":{
									"id":{"type":"integer","format":"int64"},
									"petId":{"type":"integer","format":"int64"},
									"quantity":{"type":"integer","format":"int32"},
									"shipDate":{"type":"string","format":"date-time"},
									"status":{"type":"string","description":"Order Status","enum":["placed","approved","delivered"]},
									"complete":{"type":"boolean","default":false}},
								"xml":{"name":"Order"}},
							"Category":{
								"type":"object",
								"properties":{
									"id":{"type":"integer","format":"int64"},
									"name":{"type":"string"}},
									"xml":{"name":"Category"}},
							"User":{
								"type":"object",
								"properties":{
									"token":{"type":"string","example":"---copy from login----"},
									"url":{"type":"string","example":"https://cdn.pixabay.com/photo/2016/03/28/12/35/cat-1285634_960_720.png"},
									"userStatus":{"type":"integer","format":"int32","description":"User Status"}}
								,"xml":{"name":"User"}},
							"Tag":{
								"type":"object",
								"properties":{
									"token":{"type":"string","example":"---copy from login ----"},
									"json":{"type":"string", "example":{"baz": "qux","foo": "bar"} },"patch":{"type":"string","example":[{ "op": "replace", "path": "/baz", "value": "boo" },{ "op": "add", "path": "/hello", "value": ["world"] },{ "op": "remove", "path": "/foo"}]}},
									"xml":{"name":"Tag"}},
							"Pet":{
								"type":"object",
								"required":["name","photoUrls"],
								"properties":{
									"username":{"type":"string","example":"Candidate31"},
									"password":{"type":"string","example":"asdasjfnfna"}},
									"xml":{"name":"Pet"}},"ApiResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"type":{"type":"string"},"message":{"type":"string"}}}},"externalDocs":{"description":"Find out more about Swagger","url":"http://swagger.io"}}